chore: move existint options to its own file
